Add files via upload
Based  on a clustering algorithm we are able to predict  which lending firms offer similar interest rates to women applicants.
Based on the data set  we were able to find  out which factors in the data set correctly influence the interest rate provided. 
Using a classification algorithm we were also able to determine that if the co signers credit score was bad and  the co signer of the loan had a bad credit  history the  interest rate  provided would be inflated rather applicants having co-signers with a good credit score and good repayment history.